refactor(dispatch): extract inline tests
